requirements.txt
boto3==1.24.28
matplotlib==3.5.3
nltk==3.7
protobuf==4.21.9
python-dotenv==0.21.0
spacy==3.4.2
torch==1.13.0
tqdm==4.64.1
transformers==4.24.0
